Add boundary data for regional legislature constituencies
This represents the provinces and metropolitan cities which form
the constituencies of the regional legislature
There may be some special cases eg. aosta or Friuli-Venezia Giulia
which are glossed over here... However the bulk should be good.
IDs have been created in fields MS_FB_PARE and MS_FB.
These should be in the form:
'country:it/region:lom'
and 'country:it/region:lom/province:va' or
'country:it/region:lom/metro:mi' respectivly
The crs should be 4326 and the file shoudl be UTF-8 encoded
There should be 107 features.