Built locally, Mostly sucessfully.
New warning which appears 12 times:
WARNING: the district città metropolitana di Roma Capitale (Q18288160) wasn't found in the boundary data for position sindaco di Roma (Q23936560)
This is is a little worrying because there are 12 cities,
so they may all be pointing to Rome in some way. However it
could also just be coincidence. I havn't managed to track it down,
So am opening PR for now.