Results of running `bundle exec build build | tee build_output.txt`
Note that included in this commit is the build_output.txt file.
We think that this is going to be useful to store and keep
up to date to help us keep track of changes in the warnings.

Here's what I think is the reason for the existing warnings:

WARNING: No :area_type_wikidata_item_id entry for regional-legislature-constituencies boundaries
(4 similar warnings)

This is because we added shapefiles for the national level plurinominal
constituencies to `boundaries/index.json` without :area_type_wikidata_item_id
as the nature of these new constituencies seemed unclear at election
time.

WARNING: the district America Meridionale (Q44050374) wasn't found in the boundary data for position deputato della Repubblica Italiana (Q18558478)
(644 similar warnings)

This is because we didn't add the Wikidata ID mappings for the
plurinominal constituencies.

WARNING: the district città metropolitana di Roma Capitale (Q18288160) wasn't found in the boundary data for position sindaco di Roma (Q23936560)
(83 similar warnings)

This is because the mayor the Commune of Rome automatically is
the mayor of the Metropolitan District of Rome. We're only paying
attention to the mayor of the commune, for simplicity, but the
query also returns roles for the metropolitan district.